refactor: fix atomic design
