+++ projects/common/src/constants/serverInfo.ts
@@ -1,9 +1,14 @@
 export const apiUrl = 'https://api.sasil.app' as const;
 
+export const googleClientId = {
+  web: '1039863031545-qeai77l9fsp4lpuvgj99q5hmlspeooij.apps.googleusercontent.com',
+  expo: '1039863031545-qeai77l9fsp4lpuvgj99q5hmlspeooij.apps.googleusercontent.com',
+  ios: '1039863031545-4in5ljvaephst13ridnjtqm32ododsho.apps.googleusercontent.com',
+  android:
+    '429457146313-mq693iuckrp5spudm138g1rnntgnjjt4.apps.googleusercontent.com',
+};
 export const appleClientId = 'com.sasil.sasilservice';
 export const appleRedirectURI =
   'https://sasil-front-git-apple-login-test-sasil.vercel.app/login';
 export const kakaoClientId = '2a129108cde2d21e8b1e24fb86e350e7';
 export const kakaoRedirectURI = 'https://sasil.app/login';
-export const googleClientId =
-  '1039863031545-qeai77l9fsp4lpuvgj99q5hmlspeooij.apps.googleusercontent.com';
