--- projects/sasil-mobile/src/components/templates/auth/AuthTemplate/index.ts
@@ -1,3 +1,0 @@
-import AuthTemplate from './AuthTemplate';
-
-export default AuthTemplate;
