+++ projects/sasil-mobile/src/routes/RootNavigator.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import LoginModalWrapped from '@/screens/LoginModal';
+import LoginModal from '@/screens/LoginModal/LoginModal';
 
 const Stack = createStackNavigator();
 
@@ -13,7 +13,7 @@ const RootNavigator = () => (
       }}
     >
       <Stack.Group screenOptions={{ presentation: 'modal' }}>
-        <Stack.Screen name="Login" component={LoginModalWrapped} />
+        <Stack.Screen name="Login" component={LoginModal} />
       </Stack.Group>
       {/* Login 후 넘어갈 Stack.Screen  */}
     </Stack.Navigator>
