+++ projects/sasil-web/src/components/molelcules/LoginButton/social/KakaoLoginButton.tsx
@@ -1,0 +1,42 @@
+import { kakaoClientId } from '@sasil/common';
+import LoginButton from '../LoginButton';
+import { login, getUser } from '../routes';
+
+/**
+ * Kakao 토큰으로 JWT 토큰을 받고 이를 통해 유저 정보 불러오는 함수
+ *
+ * @param authValue : Kakao로부터 받은 토큰
+ */
+// TODO : login API 연결 !!
+const responseKakao = async (authValue: string) => {
+  const res = await login(authValue, 'kakao');
+  const token = res?.data.token;
+  const user = await getUser(token);
+  return user;
+};
+
+// Kakao 로그인 함수 : 카카오 로그인 창이 열린 후, 로그인 성공시 유저 데이터를 받아오는 로직 실행
+const loginWithKakao = () => {
+  const { Kakao }: any = window;
+  Kakao.init(kakaoClientId);
+  if (Kakao.isInitialized()) {
+    Kakao.Auth.login({
+      scope: 'profile_nickname, account_email',
+      async success(authObj: any) {
+        const authValue = authObj.access_token;
+        const result = await responseKakao(authValue);
+        console.log(result);
+      },
+      fail(error: any) {
+        console.log(JSON.stringify(error));
+      },
+    });
+  }
+};
+
+// Kakao 로그인 버튼 컴포넌트
+const KakaoLoginButton = () => (
+  <LoginButton social="kakao" onClick={loginWithKakao} />
+);
+
+export default KakaoLoginButton;
