--- projects/sasil-web/src/components/templates/auth/AuthTemplate/AuthTemplate.style.tsx
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-// eslint-disable-next-line import/prefer-default-export
-export const Container = styled.div({
-  height: '100vh',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
