--- projects/sasil-web/src/components/templates/auth/AuthTemplate/index.ts
@@ -1,3 +1,0 @@
-import AuthTemplateWrapped from './AuthTemplateWrapped';
-
-export default AuthTemplateWrapped;
